add navbar, signin, signup

+++ backend/src/controllers/notes.ts
@@ -3,14 +3,22 @@ import asyncHandler from '../middlewares/asyncHandler';
 import { RequestHandler } from 'express';
 import createHttpError from 'http-errors';
 import mongoose from 'mongoose';
+import { assertIsDefined } from '../utils/assertIsDefined';
 
 export const getNotes: RequestHandler = asyncHandler(async (req, res, next) => {
-    const notes = await NoteModel.find().exec();
+    const authenticatedUserId = req.session.userId;
+
+    assertIsDefined(authenticatedUserId);
+
+    const notes = await NoteModel.find({ userId: authenticatedUserId }).exec();
     res.status(200).json({ success: true, data: notes });
 });
 
 export const getNote: RequestHandler = asyncHandler(async (req, res, next) => {
     const noteId = req.params.id;
+    const authenticatedUserId = req.session.userId;
+
+    assertIsDefined(authenticatedUserId);
 
     if (!mongoose.isValidObjectId(noteId)) {
         throw createHttpError(400, 'Invalid note id');
@@ -19,7 +27,14 @@ export const getNote: RequestHandler = asyncHandler(async (req, res, next) => {
     const note = await NoteModel.findById(noteId);
 
     if (!note) {
-        throw createHttpError(403, 'Note not found');
+        throw createHttpError(404, 'Note not found');
+    }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+        throw createHttpError(
+            401,
+            "You don'n have permission to access this note",
+        );
     }
 
     res.status(200).json({ success: true, data: note });
@@ -33,12 +48,19 @@ interface CreateNoteBody {
 export const createNote: RequestHandler<any, any, CreateNoteBody, any> =
     asyncHandler(async (req, res, next) => {
         const { title, text } = req.body;
+        const authenticatedUserId = req.session.userId;
+
+        assertIsDefined(authenticatedUserId);
 
         if (!title) {
             throw createHttpError(400, 'Note need a title');
         }
 
-        const newNote = await NoteModel.create({ title, text });
+        const newNote = await NoteModel.create({
+            userId: authenticatedUserId,
+            title,
+            text,
+        });
 
         res.status(201).json({ success: true, data: newNote });
     });
@@ -47,6 +69,9 @@ export const updateNode: RequestHandler<any, any, CreateNoteBody, any> =
     asyncHandler(async (req, res, next) => {
         const noteId = req.params.id;
         const { title, text } = req.body;
+        const authenticatedUserId = req.session.userId;
+
+        assertIsDefined(authenticatedUserId);
 
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, 'Invalid note id');
@@ -55,7 +80,14 @@ export const updateNode: RequestHandler<any, any, CreateNoteBody, any> =
         const note = await NoteModel.findById(noteId);
 
         if (!note) {
-            throw createHttpError('403', 'Note not found');
+            throw createHttpError(404, 'Note not found');
+        }
+
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(
+                401,
+                "You don'n have permission to access this note",
+            );
         }
 
         note.title = title || note.title;
@@ -69,6 +101,9 @@ export const updateNode: RequestHandler<any, any, CreateNoteBody, any> =
 export const deleteNote: RequestHandler = asyncHandler(
     async (req, res, next) => {
         const noteId = req.params.id;
+        const authenticatedUserId = req.session.userId;
+
+        assertIsDefined(authenticatedUserId);
 
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, 'Invalid note id');
@@ -77,7 +112,14 @@ export const deleteNote: RequestHandler = asyncHandler(
         const note = await NoteModel.findById(noteId);
 
         if (!note) {
-            throw createHttpError('403', 'Note not found');
+            throw createHttpError(404, 'Note not found');
+        }
+
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(
+                401,
+                "You don'n have permission to access this note",
+            );
         }
 
         const oldNote = await note.deleteOne();
